Move 2 StormEurope unittest
Tests moved:

1) climada/hazard/test/test_storm_europe/test_icon_read (~ 4.5 sec)
2 ) climada/hazard/test/test_storm_europe/test_from_footprints (~ 6 sec)

Destination:

1) climada/test/test_hazard/test_icon_read
2) climada/test/test_hazard/test_from_footprints
