+++ climada/hazard/test/test_storm_europe.py
@@ -44,55 +44,6 @@ class TestReader(unittest.TestCase):
         self.assertTrue(isinstance(cent, Centroids))
         self.assertEqual(cent.size, 9944)
 
-    def test_from_footprints(self):
-        """Test from_footprints constructor, using one small test files"""
-        def _test_first(haz):
-            """Test the expected first entry of the hazard"""
-            self.assertEqual(haz.tag.haz_type, 'WS')
-            self.assertEqual(haz.units, 'm/s')
-            self.assertEqual(haz.event_id.size, 1)
-            self.assertEqual(haz.date.size, 1)
-            self.assertEqual(dt.datetime.fromordinal(haz.date[0]).year, 1999)
-            self.assertEqual(dt.datetime.fromordinal(haz.date[0]).month, 12)
-            self.assertEqual(dt.datetime.fromordinal(haz.date[0]).day, 26)
-            self.assertEqual(haz.event_id[0], 1)
-            self.assertEqual(haz.event_name[0], 'Lothar')
-            self.assertIsInstance(haz.intensity,
-                                  sparse.csr.csr_matrix)
-            self.assertIsInstance(haz.fraction,
-                                  sparse.csr.csr_matrix)
-            self.assertEqual(haz.intensity.shape, (1, 9944))
-            self.assertEqual(haz.fraction.shape, (1, 9944))
-            self.assertEqual(haz.frequency[0], 1.0)
-
-        # Load first entry
-        storms = StormEurope.from_footprints(
-            WS_DEMO_NC[0], description='test_description')
-        _test_first(storms)
-
-        # Omit the second file, should be the same result
-        storms = StormEurope.from_footprints(WS_DEMO_NC, files_omit=str(WS_DEMO_NC[1]))
-        _test_first(storms)
-
-        # Now load both
-        storms = StormEurope.from_footprints(WS_DEMO_NC, description='test_description')
-
-        self.assertEqual(storms.tag.haz_type, 'WS')
-        self.assertEqual(storms.units, 'm/s')
-        self.assertEqual(storms.event_id.size, 2)
-        self.assertEqual(storms.date.size, 2)
-        self.assertEqual(dt.datetime.fromordinal(storms.date[0]).year, 1999)
-        self.assertEqual(dt.datetime.fromordinal(storms.date[0]).month, 12)
-        self.assertEqual(dt.datetime.fromordinal(storms.date[0]).day, 26)
-        self.assertEqual(storms.event_id[0], 1)
-        self.assertEqual(storms.event_name[0], 'Lothar')
-        self.assertIsInstance(storms.intensity,
-                              sparse.csr.csr_matrix)
-        self.assertIsInstance(storms.fraction,
-                              sparse.csr.csr_matrix)
-        self.assertEqual(storms.intensity.shape, (2, 9944))
-        self.assertEqual(storms.fraction.shape, (2, 9944))
-
     def test_read_with_ref(self):
         """Test from_footprints while passing in a reference raster."""
         storms = StormEurope.from_footprints(WS_DEMO_NC, ref_raster=WS_DEMO_NC[1])
@@ -196,47 +147,6 @@ class TestReader(unittest.TestCase):
         self.assertEqual(haz.intensity.shape, (21, 25))
         self.assertAlmostEqual(haz.intensity.max(), 36.426735,places=3)
         self.assertEqual(haz.fraction.shape, (21, 25))
-
-    def test_icon_read(self):
-        """test reading from icon grib"""
-        # for this test the forecast file is supposed to be already downloaded from the dwd
-        # another download would fail because the files are available for 24h only
-        # instead, we download it as a test dataset through the climada data api
-        apiclient = Client()
-        ds = apiclient.get_dataset_info(name='test_storm_europe_icon_2021012800', status='test_dataset')
-        dsdir, _ = apiclient.download_dataset(ds)
-        haz = StormEurope.from_icon_grib(
-            dt.datetime(2021, 1, 28),
-            dt.datetime(2021, 1, 28),
-            model_name='test',
-            grib_dir=dsdir,
-            delete_raw_data=False)
-        self.assertEqual(haz.tag.haz_type, 'WS')
-        self.assertEqual(haz.units, 'm/s')
-        self.assertEqual(haz.event_id.size, 40)
-        self.assertEqual(haz.date.size, 40)
-        self.assertEqual(dt.datetime.fromordinal(haz.date[0]).year, 2021)
-        self.assertEqual(dt.datetime.fromordinal(haz.date[0]).month, 1)
-        self.assertEqual(dt.datetime.fromordinal(haz.date[0]).day, 28)
-        self.assertEqual(haz.event_id[-1], 40)
-        self.assertEqual(haz.event_name[-1], '2021-01-28_ens40')
-        self.assertIsInstance(haz.intensity,
-                              sparse.csr.csr_matrix)
-        self.assertIsInstance(haz.fraction,
-                              sparse.csr.csr_matrix)
-        self.assertEqual(haz.intensity.shape, (40, 49))
-        self.assertAlmostEqual(haz.intensity.max(), 17.276321,places=3)
-        self.assertEqual(haz.fraction.shape, (40, 49))
-        with self.assertLogs('climada.hazard.storm_europe', level='WARNING') as cm:
-            with self.assertRaises(ValueError):
-                haz = StormEurope.from_icon_grib(
-                    dt.datetime(2021, 1, 28, 6),
-                    dt.datetime(2021, 1, 28),
-                    model_name='test',
-                    grib_dir=CONFIG.hazard.test_data.str(),
-                    delete_raw_data=False)
-        self.assertEqual(len(cm.output), 1)
-        self.assertIn('event definition is inaccuratly implemented', cm.output[0])
 
     def test_generate_forecast(self):
         """ testing generating a forecast """
